% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data.R
\docType{data}
\name{builds}
\alias{builds}
\title{builds}
\format{
A data frame with 57 rows and 4 variables:
\describe{
\item{build}{A character vector that represents the build version. It is denoted in the format \code{major.minor.hotfix}.}
\item{release_length}{An integer vector that represents the length of the release in days.}
\item{release_start}{A Date vector that represents the start date of the release.}
\item{release_end}{A Date vector that represents the end date of the release.}
}
}
\usage{
builds
}
\description{
A dataset containing information about the builds.
}
\details{
This dataset is used to track the release information of different build versions. It can be used for analyses such as release cycle time, build version impact on user activity, and release scheduling.
}
\keyword{datasets}
