\name{cohortDetails}
\alias{cohortDetails}
\title{Estimates cohort retention.}
\usage{
  cohortDetails(students, grads,
    studentIdColumn = "CONTACT_ID_SEQ",
    degreeColumn = "DEGREE_CODE",
    persistColumn = "PERSIST_FLAG",
    warehouseDateColumn = "CREATED_DATE",
    gradColumn = "START_DATE", min.cell.size = 5)
}
\description{
  Estimates cohort retention.
}

