\name{cohortRetention}
\alias{cohortRetention}
\title{Estimates cohort retention.}
\usage{
  cohortRetention(students, grads,
    studentIdColumn = "CONTACT_ID_SEQ",
    degreeColumn = "DEGREE_CODE",
    persistColumn = "PERSIST_FLAG",
    warehouseDateColumn = "CREATED_DATE",
    gradColumn = "START_DATE", grouping = NULL, ...)
}
\description{
  Estimates cohort retention.
}

