% Generated by roxygen2 (4.0.1): do not edit by hand
\name{diff.month}
\alias{diff.month}
\title{Calculates the difference in months between the two dates.}
\usage{
\method{diff}{month}(d1, d2)
}
\arguments{
\item{d1}{date one.}

\item{d2}{date two}
}
\description{
Calculates the difference in months between the two dates.
}

