% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_activity.R
\name{get_activity}
\alias{get_activity}
\title{Generate user activity data}
\usage{
get_activity(builds, users)
}
\arguments{
\item{builds}{A dataframe with build release information.}

\item{users}{A dataframe with user information.}
}
\value{
A dataframe with user activity dates and activity indicators.
}
\description{
This function generates a dataframe where each row is a combination of a row from \code{users} and a row from \code{activity_dates}.
It also calculates the \code{days_from_start} and simulates whether the user was active on a particular date, using the \link{get_activity_probability} function, dependent on number of days of activity, so that there is a much lower chance of activity the longer it has been from the user's start date.
The result is filtered so that \code{days_from_start} begins at 0.
}
\examples{
\dontrun{
builds <- get_versions() \%>\% set_build_releases()
users <- get_users(builds)
user_activity_data <- get_activity(builds, users)
}
}
