% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_activity_probability.R
\name{get_activity_probability}
\alias{get_activity_probability}
\title{Get the probability of a player returning on a given day}
\usage{
get_activity_probability(day)
}
\arguments{
\item{day}{The day from start, indexed from 0, for which to calculate the probability (integer).}
}
\value{
The probability of a player returning on the given day (numeric).
}
\description{
This function calculates the probability of a player returning on a given day.
The probability decreases over time according to the following schedule:
\itemize{
\item 100\% on day 0
\item 30\% on day 1
\item Linear decrease from 30\% to 10\% from day 1 to day 3
\item Linear decrease from 10\% to 5\% from day 3 to day 7
\item Linear decrease from 5\% to 2\% from day 7 to day 14
\item Linear decrease from 2\% to 0.05\% from day 14 to day 30
\item 0.05\% on all days after day 30
}
}
\examples{
get_activity_probability(0)  # 1
get_activity_probability(1)  # 0.3
get_activity_probability(3)  # 0.1
get_activity_probability(7)  # 0.05
get_activity_probability(14) # 0.02
get_activity_probability(30) # 0.0005
get_activity_probability(31) # 0.0005
}
