% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_builds.R
\name{get_builds}
\alias{get_builds}
\title{Generate build numbers}
\arguments{
\item{major_change_max}{The maximum value for major changes. Default is 1.}

\item{minor_change_max}{The maximum value for minor changes. Default is 1.}

\item{hot_fix_max}{The maximum value for hotfix changes. Default is 1.}
}
\value{
A data frame with the build numbers.
}
\description{
This function generates build numbers based on the maximum values for major, minor, and hotfix changes.
}
\examples{
get_builds(1, 1, 2)
}
