% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users.R
\name{get_users}
\alias{get_users}
\title{Get users from builds}
\usage{
get_users(builds, new_users_max = 2, max_activity_days = 5)
}
\arguments{
\item{builds}{A data frame of builds. This should be the output of the \code{set_build_releases} function and should include a \code{new_users} column.}

\item{new_users_max}{The maximum number of new users. This is used to generate a random number of new users for each build. Default is 2.}

\item{max_activity_days}{The maximum number of days for which to generate activity data. Default is 30.}
}
\value{
A data frame with user IDs, the build they belong to, their activity start date, and the number of activity days for each new user in each build.
}
\description{
This function filters out builds with no new users, calculates cumulative sums of new users,
generates user IDs, and unnests the user IDs into a single column. It also assigns each user a random activity start date
between the release start and end dates of the build. Defaults to a maximum of 2 new users per build.
}
\examples{
builds <- get_versions()
builds <- set_build_releases(builds, release_length_max = 30)
get_users(builds)

}
