% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_versions.R
\name{get_versions}
\alias{get_versions}
\title{Generate build numbers}
\usage{
get_versions(major_change_max = 1, minor_change_max = 1, hot_fix_max = 1)
}
\arguments{
\item{major_change_max}{The maximum value for major changes. Default is 1.}

\item{minor_change_max}{The maximum value for minor changes. Default is 1.}

\item{hot_fix_max}{The maximum value for hotfix changes. Default is 1.}
}
\value{
A data frame with the build numbers.
}
\description{
This function generates build numbers based on the maximum values for major, minor, and hotfix changes.
}
\examples{
get_versions(1, 1, 2)
}
