\docType{data}
\name{graduates}
\alias{graduates}
\title{Sample graduates data. These data do not represent actual data and
and should not be used for any inferential analysis.}
\format{a data frame 155,984 records of 3 variables}
\source{
  example graduates data
}
\description{
  Sample graduates data. These data do not represent actual
  data and and should not be used for any inferential
  analysis.
}
\keyword{datasets}

