% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.CohortRetention}
\alias{plot.CohortRetention}
\title{Plots the results of \code{\link{cohortRetention}} using \code{ggplot2}.}
\usage{
\method{plot}{CohortRetention}(cohortRetention, labelPoints = FALSE,
  textsize = 3, reverse = TRUE, title = "Cohort Based Retention",
  xlab = "Cohort", ylab1 = "Percentage", ylab2 = "Cohort Size",
  legend.position = c(0.1, 0.7), legend.justification = "left",
  retentionMonths = c(15), completionMonths = c(36, 48, 72, 96),
  plot.histogram = TRUE, useGridExtra = TRUE, ...)
}
\arguments{
\item{cohortRetention}{the results from \code{\link{cohortRetention}}.}

\item{labelPoints}{if TRUE, points will be labeled.}

\item{textsize}{the size of the text.}

\item{reverse}{if TRUE, the x-axis is reversed so most recent cohort is on the left.}

\item{title}{the plot title.}

\item{xlab}{the x-axis title.}

\item{ylab2}{the y-axis title for the histogram of cohort sizes.}

\item{ylab1}{the y-axis title for the rates.}

\item{legend.position}{the placement for the legend.}

\item{legend.justification}{the legend justification.}

\item{retentionMonths}{the months to label for the retention rates.}

\item{completionMonths}{the months to label for completion rates.}

\item{plot.histogram}{if TRUE, a histogram will be included representing cohort sizes.}

\item{useGridExtra}{if TRUE, layout options from the \code{GridExtra} package will be used.}

\item{...}{currently unused.}
}
\description{
Plots the results of \code{\link{cohortRetention}} using \code{ggplot2}.
}

