% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.Retention}
\alias{plot.Retention}
\title{Plots the results of \code{\link{retention}} use \code{ggplot2}.}
\usage{
\method{plot}{Retention}(ret, retentionMonths = c(15),
  completionMonths = c(36, 48, 72, 96), ...)
}
\arguments{
\item{ret}{the results of \code{\link{retention}}.}

\item{retentionMonths}{the months to label for the retention rates.}

\item{completionMonths}{the months to label for completion rates.}

\item{...}{currently unused.}
}
\description{
Plots the results of \code{\link{retention}} use \code{ggplot2}.
}

