% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotPersistence}
\alias{plotPersistence}
\title{Plots retention, completion, and persistence rates.}
\usage{
plotPersistence(ret)
}
\arguments{
\item{ret}{the results from \code{\link{retention}}.}
}
\description{
Plots retention, completion, and persistence rates.
}

