\name{print.CohortRetention}
\alias{print.CohortRetention}
\title{Prints the summary portion of the results}
\usage{
  print.CohortRetention(x, ...)
}
\arguments{
  \item{x}{object with class type CohortRetention.}

  \item{...}{unused parameters.}
}
\description{
  Prints the summary portion of the results
}

