% Generated by roxygen2 (4.0.1): do not edit by hand
\name{quarterlySummary}
\alias{quarterlySummary}
\title{Aggregates to quarters (i.e. three-months).}
\usage{
quarterlySummary(ret, rows, grouping = NULL, rateCol = "GraduationRate",
  firstMonth = 7)
}
\description{
Aggregates to quarters (i.e. three-months).
}

