% Generated by roxygen2 (4.0.1): do not edit by hand
\name{quarterSummaryLaTeX}
\alias{quarterSummaryLaTeX}
\title{Prints LaTeX of a quarterly aggregate.}
\usage{
quarterSummaryLaTeX(quartersum, caption = NULL, label = "quartersum",
  numQuarters = 8)
}
\description{
Prints LaTeX of a quarterly aggregate.
}

