\docType{package}
\name{retention-package}
\alias{retention}
\alias{retention-package}
\title{Retention and completion rates}
\description{
  Functions to estimate and visualize retention and
  complete rates.
}
\author{
  \email{jason@bryer.org}
}
\keyword{completion}
\keyword{institutional}
\keyword{package}
\keyword{research}
\keyword{retention}

