% Generated by roxygen2 (4.0.1): do not edit by hand
\name{retention2}
\alias{retention2}
\title{Alternative (and possibly faster) retention calculation.}
\usage{
retention2(students, graduates, months = c(15, 24, 36, 48, 72, 96),
  warehouseDateColumn = "CREATED_DATE", gradColumn = "START_DATE",
  studentIdColumn = "CONTACT_ID_SEQ", degreeColumn = "DEGREE_CODE",
  persistColumn = "Persist")
}
\description{
Alternative (and possibly faster) retention calculation.
}

