% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_build_releases.R
\name{set_build_releases}
\alias{set_build_releases}
\title{Set build releases}
\usage{
set_build_releases(builds, release_length_max = 7)
}
\arguments{
\item{builds}{A data frame of builds generated by the \code{get_builds} function.}

\item{release_length_max}{The maximum release length. This is used to generate a random release length for each build.}
}
\value{
A data frame with the build numbers, new users, release dates, and release lengths.
}
\description{
This function generates new users and release lengths, calculates cumulative release lengths,
and sets release dates for each build. Defaults to maximum length of 7 days.
}
\examples{
builds <- get_versions()
set_build_releases(builds, release_length_max = 30) 
}
