\name{studentDetails}
\alias{studentDetails}
\title{Returns student status at the specified months}
\usage{
  studentDetails(students, grads, months = c(15, 24),
    warehouseDateColumn = "CREATED_DATE",
    gradColumn = "START_DATE", ...)
}
\description{
  Returns student status at the specified months
}

