% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.Retention}
\alias{summary.Retention}
\title{Prints summary information aggregated across cohorts.}
\usage{
\method{summary}{Retention}(object, retentionMonths = c(15),
  completionMonths = c(36, 48, 72, 96), ...)
}
\arguments{
\item{retentionMonths}{the months to label for the retention rates.}

\item{completionMonths}{the months to label for completion rates.}

\item{...}{currently unused.}
}
\description{
Prints summary information aggregated across cohorts.
}

