% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data.R
\docType{data}
\name{user_activity}
\alias{user_activity}
\title{user_activity}
\format{
A data frame with 146,463 rows and 3 variables:
\describe{
\item{user}{A character vector that represents the user ID. Each user is denoted as \code{user_x}, where \code{x} is a unique identifier for each user.}
\item{build}{A character vector that represents the build version that the user is on. It is denoted in the format \code{major.minor.hotfix}.}
\item{activity_date}{A Date vector that represents the date of the user's activity.}
}
}
\usage{
user_activity
}
\description{
A dataset containing information about the activity of users.
}
\details{
This dataset is used to track the activity of users across different build versions and dates. It can be used for analyses such as user retention, activity patterns, and the impact of different build versions on user activity.
}
\keyword{datasets}
