% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data.R
\docType{data}
\name{users}
\alias{users}
\title{users}
\format{
A data frame with 47,031 rows and 4 variables:
\describe{
\item{user}{A character vector that represents the user ID. Each user is denoted as \code{user_x}, where \code{x} is a unique identifier for each user.}
\item{first_build}{A character vector that represents the first build version that the user was on. It is denoted in the format \code{major.minor.hotfix}.}
\item{activity_start}{A Date vector that represents the start date of the user's activity.}
\item{activity_days}{An integer vector that represents the number of days the user has been active.}
}
}
\usage{
users
}
\description{
A dataset containing information about the users.
}
\details{
This dataset is used to track the activity of users from their first build version and date. It can be used for analyses such as user retention, activity patterns, and the impact of different build versions on user activity.
}
\keyword{datasets}
