\name{xtable.CohortRetention}
\alias{xtable.CohortRetention}
\title{Prints LaTeX of cohort summary.}
\usage{
  xtable.CohortRetention(x, caption = NULL, label = NULL,
    align = NULL, digits = NULL, display = NULL,
    sparkline.retention.rect = NULL,
    sparkline.completion.rect = NULL, sparkline.width = 10,
    sparkline.height = NULL, retentionMonths = c(15),
    completionMonths = c(36, 48, 72, 96),
    min.cell.size = 20, ...)
}
\description{
  Prints LaTeX of cohort summary.
}

