% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.CohortRetention}
\alias{print.CohortRetention}
\title{Prints the summary portion of the results}
\usage{
\method{print}{CohortRetention}(x, ...)
}
\arguments{
\item{x}{object with class type CohortRetention.}

\item{...}{unused parameters.}
}
\description{
Prints the summary portion of the results
}

