% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.CohortRetention}
\alias{summary.CohortRetention}
\title{Prints summary of \code{\link{cohortRetention}}.}
\usage{
\method{summary}{CohortRetention}(object, months = c(15, 36, 48, 72, 96), ...)
}
\arguments{
\item{months}{the months to include in the summary.}

\item{...}{currenlty unused.}
}
\description{
Prints summary of \code{\link{cohortRetention}}.
}

